Preview Manager
Moved Preview functionality of BuildManager to PreviewManager
