+++ Assets/Scripts/HexGrid/PreviewManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a039229b65bb6754891d37b4b025ec3d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
